requirements.txt

python-oracledb>=2.4.0
# If you prefer legacy:
# cx_Oracle>=8.3.0
